pyproject.toml
[tool.poetry]
name = "tasty-link"
version = "0.1.0"
description = ""
authors = ["Dmitrii <lagodnov@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = {extras = ["standart"], version = "^0.29.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
